FONM-17496 - fix notification issues
In Exo, when the Exo-DownloadService is running, the notifications is displayed periodically.
In BellFibe, because the download queue is managed internally and not in Exo, the Exo-DownloadService continue to run a few seconds to allow us to feed with new downloads, without stopping. We do not want to display notification during those seconds, as there is no ongoing download.