Improve AV desync error detection
We added some error detection to debug an AV desync issue in trials.
We initially hard-coded the thresholds. One of them was triggering in integration tests, so I think it's better if they can be customized.
As I added the variable thresholds, I found out why it might have been triggered during integration tests and fixed it. There was no check that we'd initialized the last render time, and the renderer can be reused, so if the first frame is skipped, the error is immediately triggered.