Add a switch to enable the distinct tunneling audio session
In #32 we added a distinct audio session for tunneled playback, to fix an issue when switching between tunneled and not tunneled playback on specific setups.

We now have playback issues on some specific setups that seem to be recent.

We adding a switch to the above fix (default off) to rule out #32 as the cause for the new issue.