Workaround av desync caused by incorrect sample rate in audio chunks
We did a workaround to avoid an av desync issue when the sample rate of the chunks falsely identifies itself as a 32kHz when it's actually 48kHz. I made the fix very specific in order to limit de scope if the hack.

Turns out we can also get other values (44.1).

This PR modifies the initial hack to force any rate that is not 48 kHz to 48 kHz.

Tested on a known problematic recording, fixes the issue.