Fix reset of lastRender and skipCount (logs + non-fatals) (#15)
When we merged all of our changes in the new ExoPlayer fork, a small
block of code was merged into the wrong code block. It results in some
of our counters not resetting properly when we render a frame, which
means we can flood the logcat with false errors with the right log
level.

This PR fixes moves back the reset where it belongs.