Add a way to not ignore performance points for resolution and frame rate (#19)
In
https://github.com/androidx/media/commit/3521ccda9be020419e966308f985fa899b083a45,
a fallback was added when there is no coverage for 720p 60 FPS. That
introduced performance issues on several device models that are now
dropping 15-40 frames / sec in Live playback (depending on the model).
That's because they used to stay at a lower resolution/framerate, and
are now trying to playback 720P 60 FPS, but they struggle to do so.
This PR will allow us to disable the fallback on the identified device
models. So they will revert to using the lower fps track.