Fix the RELEASENOTES.md
This [commit](https://github.com/androidx/media/commit/d175223cc632f9a927a0cfe62bc808d336a8c83a) was introduced not long ago (after 1.4.0-alpha01), but has the release note added to the alpha01 release. Moving this piece of note to "unreleased changes" so that it will be included in the 1.4.0-alpha02 release note as a new change.

PiperOrigin-RevId: 639768495