Fixes early transiton to `Player.STATE_ENDED` with MPD transition  dynamic to static
This checkin fixes Issue #1441 where the player transitons to `Player.STATE_ENDED` once
the buffer runs out on a DASH start-over playlist that has transitioned from dynamic to static.

The fix detects the DASH Period has changed offset to the window, which occurs when the
origin vendor follows the DASH-IF recommendations in
*Section 4.6.4. Transition Phase between Live and On-Demand* ,  summerized as:

-	adds the attribute `MPD@mediaPresentationDuration`
-	removes the attribute `MPD@minimumUpdatePeriod`
-	`Period@start` is removed (if it was present)
-	`Period@duration` is added (in case more than 1 period is present)
-	`Adaptationset .SegmentTemplate@presentationTimeOffset` is set to earliest presentation time of a segment in the Adaptationset

The MPD change does not affect the render position or the segment timeline at all, however
the cleanest way to implement this was to report a `PositionUpdateForPlaylistChange`
which triggers a seek and flushes the current buffered content.